Don't delay running Python inside a timeout
As far as I can tell, this is only necessary at first load, since
the exec macro may run before on the @onload macro.  So we'll just
add an extra check that window.runPython exists before trying to
call it.

This means that we can move the setting of the pyodide_running flag
inside the runPython function, so it can be responsible for managing
that state.