Improve error reporting
In the eval mode, we can add markers to the editor indicating which
line(s) caused the problems.  To get this information from Python,
we define a custom function that looks up the last error information.
The documentation suggests we should get it from sys.exc_info(), but
this only works within the context of the error.  We're looking for
the error after it's bubbled all the way up, so we use the deprecated
sys.last_type, sys.last_value, and sys.last_traceback properties.

In exec mode, we can't do this, but to improve the display, we can
wrap the error message in <pre> tags.  Unfortunately, this doesn't
work with the second argument set to false, indicating an error (???).
So we'll set the output to be red, as would have happened with the
error flag.