Flush stdout and stderr at end of Python execution
Otherwise, output in one of those will stick around and show up
in the next cell run.  The equivalent of stream for errors is used
for writing stderr in the eval case.

In the exec case, we're still not getting console logs with writes
that are missing a trailing new line.  But they're not showing up
at the beginnings of others, so that's something I guess.  I don't
know where they're vanishing to.